chore: remove usage of SyntaxRewriter
This commit changes the architecture from using Slang's SyntaxRewriter to
using a simple text based regex expansion.  Trying to deal with the
leading trivia model of Slang is too painful.  Instead, the AST is
parsed to identify all source locations and build the required structs,
followed by a second pass that does the string replacements.